// src/app/interfaces/ticket.interface.ts
// {
//   "subject": "I have a question",
//   "name": "John Doe",
//   "email": "lawrence.fan@whatever.com",
//   "phone": "123-456-7890",
//   "message": "I have a question about my order"
// }

export interface Ticket {
    subject: string;
    name: string;
    email: string;
    phone: string;
    message: string;
}
    